+++ lib/src/transformers/group_by.dart
@@ -1,88 +1,151 @@
 import 'dart:async';
 
-class _GroupByStreamSink<S, T> implements EventSink<S> {
-  final T Function(S event) _grouper;
-  final EventSink<GroupByStream<S, T>> _outputSink;
-  final _mapper = <T, StreamController<S>>{};
+import 'package:rxdart/src/utils/forwarding_sink.dart';
+import 'package:rxdart/src/utils/forwarding_stream.dart';
 
-  _GroupByStreamSink(this._outputSink, this._grouper);
+class _GroupByStreamSink<T, K> extends ForwardingSink<T, GroupedStream<T, K>> {
+  final K Function(T event) grouper;
+  final Stream<void> Function(GroupedStream<T, K>)? duration;
 
-  @override
-  void add(S data) {
-    final key = _grouper(data);
+  final groups = <K, StreamController<T>>{};
+  Map<K, StreamSubscription<void>>? subscriptions;
 
-    final groupedController =
-        _mapper.putIfAbsent(key, () => _controllerBuilder(key));
+  _GroupByStreamSink(this.grouper, this.duration);
 
-    groupedController.add(data);
+  void _closeAll() {
+    groups.values.forEach((c) => c.close());
+    groups.clear();
   }
 
-  @override
-  void addError(e, [st]) => _outputSink.addError(e, st);
+  StreamController<T> _controllerBuilder(K key) {
+    final groupedController = StreamController<T>.broadcast(sync: true);
+    final groupByStream = GroupedStream<T, K>(key, groupedController.stream);
 
-  @override
-  void close() {
-    _mapper.values.forEach((c) => c.close());
-    _mapper.clear();
+    if (duration != null) {
+      subscriptions?.remove(key)?.cancel();
+      (subscriptions ??= {})[key] = duration!(groupByStream).take(1).listen(
+        null,
+        onDone: () {
+          subscriptions!.remove(key);
+          groups.remove(key)?.close();
+        },
+        onError: onError,
+      );
+    }
 
-    _outputSink.close();
-  }
-
-  StreamController<S> _controllerBuilder(T forKey) {
-    final groupedController = StreamController<S>();
-
-    _outputSink.add(GroupByStream<S, T>(forKey, groupedController.stream));
-
+    sink.add(groupByStream);
     return groupedController;
   }
+
+  @override
+  void onData(T data) {
+    final K key;
+    try {
+      key = grouper(data);
+    } catch (e, s) {
+      sink.addError(e, s);
+      return;
+    }
+
+    groups.putIfAbsent(key, () => _controllerBuilder(key)).add(data);
+  }
+
+  @override
+  void onError(e, st) => sink.addError(e, st);
+
+  @override
+  void onDone() {
+    _closeAll();
+    sink.close();
+  }
+
+  @override
+  FutureOr<void> onCancel() {
+    scheduleMicrotask(_closeAll);
+
+    if (subscriptions?.isNotEmpty == true) {
+      final future = Future.wait(subscriptions!.values.map((s) => s.cancel()));
+      subscriptions?.clear();
+      subscriptions = null;
+      return future;
+    } else {
+      return null;
+    }
+  }
+
+  @override
+  FutureOr<void> onListen() {}
+
+  @override
+  void onPause() => subscriptions?.values.forEach((s) => s.pause());
+
+  @override
+  void onResume() => subscriptions?.values.forEach((s) => s.resume());
 }
 
 /// The GroupBy operator divides a [Stream] that emits items into
-/// a [Stream] that emits [GroupByStream],
+/// a [Stream] that emits [GroupedStream],
 /// each one of which emits some subset of the items
 /// from the original source [Stream].
 ///
-/// [GroupByStream] acts like a regular [Stream], yet
+/// [GroupedStream] acts like a regular [Stream], yet
 /// adding a 'key' property, which receives its [Type] and value from
 /// the [_grouper] Function.
 ///
-/// All items with the same key are emitted by the same [GroupByStream].
-class GroupByStreamTransformer<S, T>
-    extends StreamTransformerBase<S, GroupByStream<S, T>> {
-  /// Method which converts incoming events into a new [GroupByStream]
-  final T Function(S event) grouper;
+/// All items with the same key are emitted by the same [GroupedStream].
+class GroupByStreamTransformer<T, K>
+    extends StreamTransformerBase<T, GroupedStream<T, K>> {
+  /// Method which converts incoming events into a new [GroupedStream]
+  final K Function(T event) grouper;
+
+  /// A function that returns an [Stream] to determine how long each group should exist.
+  /// When the returned [Stream] emits its first data or done event,
+  /// the group will be closed and removed.
+  final Stream<void> Function(GroupedStream<T, K> grouped)? durationSelector;
 
   /// Constructs a [StreamTransformer] which groups events from the source
-  /// [Stream] and emits them as [GroupByStream].
-  GroupByStreamTransformer(this.grouper);
+  /// [Stream] and emits them as [GroupedStream].
+  GroupByStreamTransformer(this.grouper, {this.durationSelector});
 
   @override
-  Stream<GroupByStream<S, T>> bind(Stream<S> stream) => Stream.eventTransformed(
-      stream, (sink) => _GroupByStreamSink<S, T>(sink, grouper));
+  Stream<GroupedStream<T, K>> bind(Stream<T> stream) => forwardStream(
+      stream, () => _GroupByStreamSink<T, K>(grouper, durationSelector));
 }
 
 /// The [Stream] used by [GroupByStreamTransformer], it contains events
 /// that are grouped by a key value.
-class GroupByStream<T, S> extends StreamView<T> {
+class GroupedStream<T, K> extends StreamView<T> {
   /// The key is the category to which all events in this group belong to.
-  final S key;
+  final K key;
 
   /// Constructs a [Stream] which only emits events that can be
   /// categorized under [key].
-  GroupByStream(this.key, Stream<T> stream) : super(stream);
+  GroupedStream(this.key, Stream<T> stream) : super(stream);
+
+  @override
+  String toString() => 'GroupedStream{key: $key}';
 }
 
 /// Extends the Stream class with the ability to convert events into Streams
 /// of events that are united by a key.
 extension GroupByExtension<T> on Stream<T> {
   /// The GroupBy operator divides a [Stream] that emits items into a [Stream]
-  /// that emits [GroupByStream], each one of which emits some subset of the
+  /// that emits [GroupedStream], each one of which emits some subset of the
   /// items from the original source [Stream].
   ///
-  /// [GroupByStream] acts like a regular [Stream], yet adding a 'key' property,
+  /// [GroupedStream] acts like a regular [Stream], yet adding a 'key' property,
   /// which receives its [Type] and value from the [grouper] Function.
   ///
-  /// All items with the same key are emitted by the same [GroupByStream].
-  Stream<GroupByStream<T, S>> groupBy<S>(S Function(T value) grouper) =>
-      transform(GroupByStreamTransformer<T, S>(grouper));
+  /// All items with the same key are emitted by the same [GroupedStream].
+  ///
+  /// Optionally, `groupBy` takes a second argument [durationSelector].
+  /// [durationSelector] is a function that returns an [Stream] to determine how long
+  /// each group should exist. When the returned [Stream] emits its first data or done event,
+  /// the group will be closed and removed.
+  Stream<GroupedStream<T, K>> groupBy<K>(
+    K Function(T value) grouper, {
+    Stream<void> Function(GroupedStream<T, K> grouped)? durationSelector,
+  }) =>
+      transform(GroupByStreamTransformer<T, K>(grouper,
+          durationSelector: durationSelector));
 }